refactor: change the time type uint64_t to uint

+++ src/dora_protocol.h
@@ -20,7 +20,7 @@ struct Peer
 {
     QString username;
     QString platform;
-    uint64_t lastUpdatedTime;
+    uint    lastUpdatedTime;
 };
 
 /**
